+++ components/com_kinoarhiv/views/movies/tmpl/default.php
@@ -206,17 +206,17 @@ JHtml::_('script', 'components/com_kinoarhiv/assets/js/jquery.lazyload.min.js');
 														<?php }
 													} ?>
 													<?php if ($this->params->get('ratings_img_kp') != 0 && !empty($item->kp_id)): ?>
-														<a href="http://www.kinopoisk.ru/film/<?php echo $item->kp_id; ?>/" rel="nofollow" target="_blank">
+														<a href="https://www.kinopoisk.ru/film/<?php echo $item->kp_id; ?>/" rel="nofollow" target="_blank">
 															<?php if ($this->params->get('ratings_img_kp_remote') == 0): ?>
 																<img src="<?php echo $rating_image_www; ?>/kinopoisk/<?php echo $item->id; ?>_big.png" border="0"/>
 															<?php else: ?>
-																<img src="http://www.kinopoisk.ru/rating/<?php echo $item->kp_id; ?>.gif" border="0" style="padding-left: 1px;"/>
+																<img src="https://www.kinopoisk.ru/rating/<?php echo $item->kp_id; ?>.gif" border="0" style="padding-left: 1px;"/>
 															<?php endif; ?>
 														</a>
 													<?php endif; ?>
 													<?php if ($this->params->get('ratings_img_rotten') != 0 && !empty($item->rottentm_id)): ?>
 														<?php if (file_exists($this->params->get('media_rating_image_root') . '/rottentomatoes/' . $item->id . '_big.png')): ?>
-															<a href="http://www.rottentomatoes.com/m/<?php echo $item->rottentm_id; ?>/" rel="nofollow" target="_blank"><img src="<?php echo $rating_image_www; ?>/rottentomatoes/<?php echo $item->id; ?>_big.png" border="0"/></a>
+															<a href="https://www.rottentomatoes.com/m/<?php echo $item->rottentm_id; ?>/" rel="nofollow" target="_blank"><img src="<?php echo $rating_image_www; ?>/rottentomatoes/<?php echo $item->id; ?>_big.png" border="0"/></a>
 														<?php endif; ?>
 													<?php endif; ?>
 													<?php if ($this->params->get('ratings_img_metacritic') != 0 && !empty($item->metacritics_id)): ?>
@@ -241,7 +241,7 @@ JHtml::_('script', 'components/com_kinoarhiv/assets/js/jquery.lazyload.min.js');
 												<?php if (!empty($item->kp_votesum) && !empty($item->kp_votes)): ?>
 													<div id="rate-kp">
 														<span class="a"><?php echo JText::_('COM_KA_RATE_KP'); ?></span>
-														<span class="b"><a href="http://www.kinopoisk.ru/film/<?php echo $item->kp_id; ?>/" rel="nofollow" target="_blank"><?php echo $item->kp_votesum; ?>
+														<span class="b"><a href="https://www.kinopoisk.ru/film/<?php echo $item->kp_id; ?>/" rel="nofollow" target="_blank"><?php echo $item->kp_votesum; ?>
 																(<?php echo $item->kp_votes; ?>)</a></span></div>
 												<?php else: ?>
 													<div id="rate-kp">
@@ -251,7 +251,7 @@ JHtml::_('script', 'components/com_kinoarhiv/assets/js/jquery.lazyload.min.js');
 												<?php if (!empty($item->rate_fc)): ?>
 													<div id="rate-rt">
 														<span class="a"><?php echo JText::_('COM_KA_RATE_RT'); ?></span>
-														<span class="b"><a href="http://www.rottentomatoes.com/m/<?php echo $item->rottentm_id; ?>/" rel="nofollow" target="_blank"><?php echo $item->rate_fc; ?>
+														<span class="b"><a href="https://www.rottentomatoes.com/m/<?php echo $item->rottentm_id; ?>/" rel="nofollow" target="_blank"><?php echo $item->rate_fc; ?>
 																%</a></span></div>
 												<?php else: ?>
 													<div id="rate-rt">