+++ components/com_kinoarhiv/views/release/tmpl/default.php
@@ -79,17 +79,17 @@ JHtml::_('script', 'components/com_kinoarhiv/assets/js/jquery.rateit.min.js');
 											<?php }
 										} ?>
 										<?php if ($this->params->get('ratings_img_kp') != 0 && !empty($this->item->kp_id)): ?>
-											<a href="http://www.kinopoisk.ru/film/<?php echo $this->item->kp_id; ?>/" rel="nofollow" target="_blank">
+											<a href="https://www.kinopoisk.ru/film/<?php echo $this->item->kp_id; ?>/" rel="nofollow" target="_blank">
 												<?php if ($this->params->get('ratings_img_kp_remote') == 0): ?>
 													<img src="<?php echo $rating_image_www; ?>/kinopoisk/<?php echo $this->item->id; ?>_big.png" border="0"/>
 												<?php else: ?>
-													<img src="http://www.kinopoisk.ru/rating/<?php echo $this->item->kp_id; ?>.gif" border="0" style="padding-left: 1px;"/>
+													<img src="https://www.kinopoisk.ru/rating/<?php echo $this->item->kp_id; ?>.gif" border="0" style="padding-left: 1px;"/>
 												<?php endif; ?>
 											</a>
 										<?php endif; ?>
 										<?php if ($this->params->get('ratings_img_rotten') != 0 && !empty($this->item->rottentm_id)): ?>
 											<?php if (file_exists($this->params->get('media_rating_image_root') . '/rottentomatoes/' . $this->item->id . '_big.png')): ?>
-												<a href="http://www.rottentomatoes.com/m/<?php echo $this->item->rottentm_id; ?>/" rel="nofollow" target="_blank"><img src="<?php echo $rating_image_www; ?>/rottentomatoes/<?php echo $this->item->id; ?>_big.png" border="0"/></a>
+												<a href="https://www.rottentomatoes.com/m/<?php echo $this->item->rottentm_id; ?>/" rel="nofollow" target="_blank"><img src="<?php echo $rating_image_www; ?>/rottentomatoes/<?php echo $this->item->id; ?>_big.png" border="0"/></a>
 											<?php endif; ?>
 										<?php endif; ?>
 										<?php if ($this->params->get('ratings_img_metacritic') != 0 && !empty($this->item->metacritics_id)): ?>
@@ -112,7 +112,7 @@ JHtml::_('script', 'components/com_kinoarhiv/assets/js/jquery.rateit.min.js');
 									<?php if (!empty($this->item->kp_votesum) && !empty($this->item->kp_votes)): ?>
 										<div id="rate-kp">
 											<span class="a"><?php echo JText::_('COM_KA_RATE_KP'); ?></span>
-											<span class="b"><a href="http://www.kinopoisk.ru/film/<?php echo $this->item->kp_id; ?>/" rel="nofollow" target="_blank"><?php echo $this->item->kp_votesum; ?>
+											<span class="b"><a href="https://www.kinopoisk.ru/film/<?php echo $this->item->kp_id; ?>/" rel="nofollow" target="_blank"><?php echo $this->item->kp_votesum; ?>
 													(<?php echo $this->item->kp_votes; ?>)</a></span></div>
 									<?php else: ?>
 										<div id="rate-kp">
@@ -122,7 +122,7 @@ JHtml::_('script', 'components/com_kinoarhiv/assets/js/jquery.rateit.min.js');
 									<?php if (!empty($this->item->rate_fc)): ?>
 										<div id="rate-rt">
 											<span class="a"><?php echo JText::_('COM_KA_RATE_RT'); ?></span>
-											<span class="b"><a href="http://www.rottentomatoes.com/m/<?php echo $this->item->rottentm_id; ?>/" rel="nofollow" target="_blank"><?php echo $this->item->rate_fc; ?>
+											<span class="b"><a href="https://www.rottentomatoes.com/m/<?php echo $this->item->rottentm_id; ?>/" rel="nofollow" target="_blank"><?php echo $this->item->rate_fc; ?>
 													%</a></span></div>
 									<?php else: ?>
 										<div id="rate-rt">